% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.function.R
\name{as.function.formula}
\alias{as.function.formula}
\title{Convert into function}
\usage{
\method{as.function}{formula}(x, ...)
}
\arguments{
\item{x}{(formula)}

\item{...}{not used}
}
\description{
Convert a formula into a function. See \link{map} for examples.
}

