% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindRows.R
\name{bindRows}
\alias{bindRows}
\title{Bind rows}
\usage{
bindRows(x, id = NULL)
}
\arguments{
\item{x}{(list)}

\item{id}{passed to \link[dplyr]{bind_rows} as \code{.id}}
}
\value{
If the first element of \code{x} inherits from \code{data.frame} the type
that first eleent.

\code{x} else.
}
\description{
This is a wrapper around \link[dplyr]{bind_rows} to preserve the input class.
}
