% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperTypes.R
\name{FL}
\alias{FL}
\title{Dynamically generate formulas}
\usage{
FL(..., .n = ".n", pattern = "\\\\.n")
}
\arguments{
\item{...}{(formulas)}

\item{.n}{(character) names to be used in formulas}

\item{pattern}{(character) pattern to be replaced in formulas}
}
\description{
Function to dynamically generate formulas to be used in \link{mutar}.
}
\examples{

FL(.n ~ mean(.n), .n = "variable")
}

