% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperTypes.R
\name{FL}
\alias{FL}
\title{Dynamically generate formulas}
\usage{
FL(..., .n = ".n", pattern = "\\\\.n", lazy = isLazyFormula(.n))
}
\arguments{
\item{...}{(formulas)}

\item{.n}{(character) names to be used in formulas}

\item{pattern}{(character) pattern to be replaced in formulas}

\item{lazy}{(logical) flag to indicate if \code{.n} should be interpreted
within the data, i.e. in the future. In this cas \code{.n} is an object
which can be used in \link{extract}.}
}
\description{
Function to dynamically generate formulas - (F)ormula (L)ist  to be used in
\link{mutar}.
}
\examples{
FL(.n ~ mean(.n), .n = "variable")
}
\seealso{
\link{mutar}
}

