% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List.R
\docType{class}
\name{List-class}
\alias{L}
\alias{List}
\alias{List-class}
\title{S4 list}
\usage{
L(...)

List(...)
}
\arguments{
\item{...}{arguments passed to \link{list}}
}
\description{
A type inheriting from list. Mainly used for dispatch in \link{map}.
}

