% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List.R
\docType{class}
\name{MList-class}
\alias{ML}
\alias{MList}
\alias{MList-class}
\title{S4 list}
\usage{
ML(...)

MList(...)
}
\arguments{
\item{...}{arguments passed to \link{list}}
}
\description{
A type inheriting from list. Used for dispatch in \link{map}.
}

