% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\docType{methods}
\name{reduce}
\alias{reduce}
\alias{reduce,ANY,formula-method}
\alias{reduce,ANY,function-method}
\title{Reduce}
\usage{
reduce(x, f, init = NULL)

\S4method{reduce}{ANY,`function`}(x, f, init = NULL)

\S4method{reduce}{ANY,formula}(x, f, init = NULL)
}
\arguments{
\item{x}{(ANY) a vector to be reduced}

\item{f}{(function | formula) a binary function}

\item{init}{an optional initial value}
}
\description{
Reduce
}

