% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\docType{methods}
\name{replace}
\alias{replace}
\alias{replace,ANY,character-method}
\alias{replace,ANY,formula-method}
\alias{replace,atomic,function-method}
\alias{replace,atomicORlist,integerORnumericORlogical-method}
\alias{replace,list,function-method}
\title{Replace elements in a vector}
\usage{
replace(x, ind, values, ...)

\S4method{replace}{list,`function`}(x, ind, values, ...)

\S4method{replace}{atomic,`function`}(x, ind, values, ...)

\S4method{replace}{atomicORlist,integerORnumericORlogical}(x, ind, values, ...)

\S4method{replace}{ANY,formula}(x, ind, values, ...)

\S4method{replace}{ANY,character}(x, ind, values, ...)
}
\arguments{
\item{x}{(atomic | list) a vector.}

\item{ind}{(function | formula | character | numeric | integer | logical) a
formula is coerced into a function. For lists the function is applied to
each element (and has to return a logical of length 1). For atomics a
vectorized function is expected. If you supply an atomic it is used for
subsetting. A character of length 1 beginning with "^" is interpreted as
regular expression.}

\item{values}{the values for replacement.}

\item{...}{arguments passed to \code{ind} if it can be interpreted as
function. For a regex arguments are passed to \link{grep}.}
}
\description{
This function replaces elements in a vector. It is a link to
\link[base]{replace} as a generic function.
}
\examples{
replace(c(1, 2, NA), is.na, 0)
replace(c(1, 2, NA), rep(TRUE, 3), 0)
replace(c(1, 2, NA), 3, 0)
replace(list(x = 1, y = 2), "x", 0)
}

