% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.function.R
\name{as.function.formula}
\alias{as.function.formula}
\title{Coerce a formula into a function}
\usage{
\method{as.function}{formula}(x, ...)
}
\arguments{
\item{x}{(formula) see examples}

\item{...}{not used}
}
\value{
An object inheriting from class function.
}
\description{
Convert a formula into a function. See \link{map} and \link{extract} for
examples.
}
\examples{
as.function(~ .)(1)
as.function(x ~ x)(1)
as.function(f(x, y) ~ c(x, y))(1, 2)
as.function(numeric : x ~ x)(1) # check for class
as.function(numeric(1) : x ~ x)(1) # check for class + length
}
