% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindRows.R
\name{bindRows}
\alias{bindRows}
\title{Bind rows}
\usage{
bindRows(x, id = NULL, useNames = TRUE, fill = TRUE)
}
\arguments{
\item{x}{(list) a list of data frames}

\item{id, useNames, fill}{passed to \link[data.table]{rbindlist}}
}
\value{
If the first element of \code{x} inherits from \code{data.frame} the type
that first element. 

\code{x} else.
}
\description{
This is a wrapper around \link[data.table]{rbindlist} to preserve the input
class.
}
