% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\name{DataFrame}
\alias{DataFrame}
\alias{[.DataFrame}
\alias{as.DataFrame}
\alias{as.DataFrame.default}
\title{DataFrame and methods}
\usage{
DataFrame(...)

as.DataFrame(x)

\method{as.DataFrame}{default}(x)

\method{[}{DataFrame}(x, i, j, ..., by, drop)
}
\arguments{
\item{...}{arbitrary number of args
\cr in \code{[} (TwoSidedFormulas)
\cr in constructor see \link[dplyr]{data_frame}}

\item{x}{(DataFrame | ANY)}

\item{i}{(logical | numeric | integer | OneSidedFormula | TwoSidedFormula)}

\item{j}{(logical | character | TwoSidedFormula)}

\item{by}{(character) variable name}

\item{drop}{ignored}
}
\description{
DataFrame and methods
}
\examples{
\dontrun{
vignette("Introduction", "dat")
}

}

