% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\name{DataFrame}
\alias{DataFrame}
\alias{[.DataFrame}
\alias{as.DataFrame}
\alias{as.DataFrame.default}
\title{DataFrame and methods}
\usage{
DataFrame(...)

as.DataFrame(x)

\method{as.DataFrame}{default}(x)

\method{[}{DataFrame}(x, i, j, ..., by, drop)
}
\arguments{
\item{...}{dots}

\item{x}{an object}

\item{i}{(logical | numeric | integer | OneSidedFormula)}

\item{j}{(logical | character | TwoSidedFormula)}

\item{by}{(character) variable name}

\item{drop}{ignored}

\item{x}{(DataFrame)}

\item{...}{(TwoSidedFormulas)}
}
\description{
DataFrame and methods
}

