% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\name{DataFrame}
\alias{DataFrame}
\alias{[.DataFrame}
\alias{as.DataFrame}
\alias{as.DataFrame.data.frame}
\alias{as.DataFrame.default}
\title{DataFrame and methods}
\usage{
DataFrame(...)

as.DataFrame(x, ...)

\method{as.DataFrame}{default}(x, ...)

\method{as.DataFrame}{data.frame}(x, ...)

\method{[}{DataFrame}(x, i, j, ..., by, sby, drop)
}
\arguments{
\item{...}{arbitrary number of args
\cr in \code{[} (TwoSidedFormulas)
\cr in constructor see \link[dplyr]{data_frame}}

\item{x}{(DataFrame | data.frame)}

\item{i}{(logical | numeric | integer | OneSidedFormula | TwoSidedFormula)}

\item{j}{(logical | character | TwoSidedFormula | function) character
beginning with '^' are interpreted as regular expression}

\item{by, sby}{(character) variable names used in \link{group_by}. Using `sby`
triggers a summarise.}

\item{drop}{ignored}
}
\description{
There are many ways to use a DataFrame. Please see the vignette to see some
examples.
}
\details{
\code{OneSidedFormula} is always used for subsetting rows.

\code{TwoSidedFormula} is used instead of name-value expressions in
\code{summarise} and \code{mutate}.
}
\examples{
\dontrun{
  vignette("Introduction", "dat")
}

}

