% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\name{DataFrame}
\alias{DataFrame}
\alias{[.DataFrame}
\alias{as.DataFrame}
\alias{as.DataFrame.data.frame}
\alias{as.DataFrame.default}
\alias{mutar}
\title{DataFrame and methods}
\usage{
DataFrame(...)

as.DataFrame(x, ...)

\method{as.DataFrame}{default}(x, ...)

\method{as.DataFrame}{data.frame}(x, ...)

\method{[}{DataFrame}(x, i, j, ..., by, drop)

mutar(x, i, j, ..., by, drop)
}
\arguments{
\item{...}{arbitrary number of args
\cr in \code{[} (TwoSidedFormulas)
\cr in constructor see \link[dplyr]{data_frame}}

\item{x}{(DataFrame | data.frame)}

\item{i}{(logical | numeric | integer | OneSidedFormula | TwoSidedFormula)}

\item{j}{(logical | character | TwoSidedFormula | function) character
beginning with '^' are interpreted as regular expression}

\item{by}{(character) variable name}

\item{drop}{ignored}
}
\description{
There are many ways to use a DataFrame. Please see the vignette to see some
examples. \code{mutar} is literally the same function as \code{[.DataFrame}
and can be used as a generic interface to use dplyr. You can use \code{mutar}
even if you do not want to use a \code{DataFrame}. Also it can be used with
S4 data.frame(s) / data_frame(s) / data.table(s).
}
\details{
\code{OneSidedFormula} is always used for subsetting rows.

\code{TwoSidedFormula} is used instead of name-value expressions in
\code{summarise} and \code{mutate}.
}
\examples{
\dontrun{
  vignette("Introduction", "dat")
}

}

