% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract,ANY,character-method}
\alias{extract,ANY,formula-method}
\alias{extract,atomic,function-method}
\alias{extract,atomicORlist,numericORintegerORlogical-method}
\alias{extract,list,function-method}
\title{Extract elements from a vector}
\usage{
extract(x, f, ...)

\S4method{extract}{list,`function`}(x, f, ...)

\S4method{extract}{atomic,`function`}(x, f, ...)

\S4method{extract}{ANY,formula}(x, f, ...)

\S4method{extract}{atomicORlist,numericORintegerORlogical}(x, f, ...)

\S4method{extract}{ANY,character}(x, f, ...)
}
\arguments{
\item{x}{(ANY) a vector}

\item{f}{(function | formula | character | numeric | integer | logical) a
formula is coerced into a function. For lists the function is applied to
each element (and has to return a logical of length 1). For atomics a
vectorized function is expected. If you supply an atomic it is used for
subsetting. A character of length 1 beginning with "^" is interpreted as
regular expression.}

\item{...}{arguments passed to f}
}
\description{
Extract elements from a vector
}
\examples{
extract(1:15, ~ 15 \%\% . == 0)
extract(list(xy = 1, zy = 2), "^z")
extract(list(x = 1, z = 2), 1)
extract(list(x = 1, y = ""), is.character)
}

