% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperTypes.R
\docType{class}
\name{MList-class}
\alias{Dots}
\alias{FunctionWithDots-class}
\alias{ML}
\alias{MList}
\alias{MList-class}
\title{Helper Types}
\usage{
ML(...)

MList(...)

Dots(f)
}
\arguments{
\item{...}{arguments passed to \link{list}}

\item{f}{(function)}
}
\description{
Defferent types used for dispatch. \code{ML} and \code{MList} are constructor
functions for a list (used in \link{map} to indicate multivariate map).
\code{Dots} wraps a function in a type to declare that it knows how to deal
with an arbitrary number of arguments (used in \link{reduce} instead of
\code{do.call}).
}

