% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperTypes.R
\docType{class}
\name{MList-class}
\alias{ML}
\alias{MList}
\alias{MList-class}
\title{Helper Types}
\usage{
ML(...)

MList(...)
}
\arguments{
\item{...}{arguments passed to \link{list}}
}
\description{
Different types used for dispatch. \code{ML} and \code{MList} are constructor
functions for a list (used in \link{map} to indicate multivariate map). A
MList is just a S4 list. You probably do not need to use this type
interactively.
}
\examples{
MList(x = 1, y = "a")
}

