% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperTypes.R
\docType{class}
\name{MList-class}
\alias{ML}
\alias{MList}
\alias{MList-class}
\title{Helper Types}
\usage{
ML(...)

MList(...)
}
\arguments{
\item{...}{arguments passed to \link{list}}
}
\description{
Helper Types
}

