% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperTypes.R
\docType{class}
\name{MList-class}
\alias{Dots}
\alias{FunctionWithDots}
\alias{FunctionWithDots-class}
\alias{ML}
\alias{MList}
\alias{MList-class}
\title{Helper Types}
\usage{
ML(...)

MList(...)

Dots(f)

FunctionWithDots(...)
}
\arguments{
\item{...}{arguments passed to \link{list}}

\item{f}{(function)}
}
\description{
Defferent types used for dispatch.
}

