% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\docType{methods}
\name{reduce}
\alias{reduce}
\alias{reduce,ANY,FunctionWithDots,missing-method}
\alias{reduce,ANY,functionORformula,missing-method}
\title{Reduce}
\usage{
reduce(x, f, init)

\S4method{reduce}{ANY,functionORformula,missing}(x, f, init)

\S4method{reduce}{ANY,FunctionWithDots,missing}(x, f, init)
}
\arguments{
\item{x}{(ANY) a vector to be reduced}

\item{f}{(function | formula) a binary function}

\item{init}{an optional initial value}
}
\description{
Very similar to \link{Reduce}.
}
\examples{
add <- function(x) reduce(x, `+`)
add(1:3)

# Standard but inefficient:
reduce(as.list(1:3), c)
# Using Dots can be more efficient when the function itself already knows
# how to reduce a vector into a scalar.
reduce(as.list(1:3), Dots(c))
# The result is the same as
do.call(c, as.list(1:3))
}

