% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\docType{methods}
\name{reduce}
\alias{reduce}
\alias{reduce,ANY,FunctionWithDots,missing-method}
\alias{reduce,ANY,functionORformula,missing-method}
\title{Reduce}
\usage{
reduce(x, f, init)

\S4method{reduce}{ANY,functionORformula,missing}(x, f, init)

\S4method{reduce}{ANY,FunctionWithDots,missing}(x, f, init)
}
\arguments{
\item{x}{(ANY) a vector to be reduced}

\item{f}{(function | formula) a binary function}

\item{init}{an optional initial value}
}
\description{
Very similar to \link{Reduce}.
}

