% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\docType{methods}
\name{replace}
\alias{replace}
\alias{replace,ANY,character-method}
\alias{replace,ANY,formula-method}
\alias{replace,ANY,logical-method}
\alias{replace,atomic,function-method}
\alias{replace,atomicORlist,integerORnumeric-method}
\alias{replace,list,function-method}
\title{Replace elements in a vector}
\usage{
replace(x, f, with, ...)

\S4method{replace}{list,`function`}(x, f, with, ...)

\S4method{replace}{atomic,`function`}(x, f, with, ...)

\S4method{replace}{atomicORlist,integerORnumeric}(x, f, with, ...)

\S4method{replace}{ANY,formula}(x, f, with, ...)

\S4method{replace}{ANY,logical}(x, f, with, ...)

\S4method{replace}{ANY,character}(x, f, with, ...)
}
\arguments{
\item{x}{(ANY) a vector}

\item{f}{(function | formula | character | numeric | integer | logical) a
formula is coerced into a function. For lists the function is applied to
each element (and has to return a logical of length 1). For atomics a
vectorized function is expected. If you supply an atomic it is used for
subsetting. A character of length 1 beginning with "^" is interpreted as
regular expression.}

\item{with}{the values for replacement}

\item{...}{arguments passed to f}
}
\description{
This function replaces elements in a vector.
}
\examples{
replace(c(1, 2, NA), is.na, 0)
replace(c(1, 2, NA), rep(TRUE, 3), 0)
replace(c(1, 2, NA), 3, 0)
replace(list(x = 1, y = 2), "x", 0)
}

