% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\docType{methods}
\name{replace}
\alias{replace}
\alias{replace,ANY,character-method}
\alias{replace,ANY,formula-method}
\alias{replace,atomic,function-method}
\alias{replace,atomicORlist,integerORnumericORlogical-method}
\alias{replace,list,function-method}
\title{Replace elements in a vector}
\usage{
replace(x, ind, values, ...)

\S4method{replace}{list,`function`}(x, ind, values, ...)

\S4method{replace}{atomic,`function`}(x, ind, values, ...)

\S4method{replace}{atomicORlist,integerORnumericORlogical}(x, ind, values, ...)

\S4method{replace}{ANY,formula}(x, ind, values, ...)

\S4method{replace}{ANY,character}(x, ind, values, ...)
}
\arguments{
\item{x}{(atomic | list) a vector.}

\item{ind}{(integer | numeric | logical | character | function | formula) a
formula is coerced to a function. A function is expected to return a
logical of length one when applied to a list; otherwise a vector which can
be used for subsetting of \code{x}. A character can be used when \code{x}
is named; when has length one and a leading \code{^} it is interpreted as
regex. Otherwise something which can be used for subsetting is expected.}

\item{values}{the values for replacement.}

\item{...}{arguments passed to \code{ind} if it can be interpreted as
function. For a regex arguments are passed to \link{grep}.}
}
\description{
This function replaces elements in a vector. It is a link to
\link[base]{replace} as a generic function.
}
\examples{
replace(c(1, 2, NA), is.na, 0)
replace(c(1, 2, NA), rep(TRUE, 3), 0)
replace(c(1, 2, NA), 3, 0)
replace(list(x = 1, y = 2), "x", 0)
}

